BUG: Fix $cdashpath for Windows to find log file correctly. Fix reporting error on deleting log file if it wasn't there in the first place...
ENH: Add ability to override config.test.php with a config.test.local.php, just like we already have for the main cdash/config.php. That way, a client running testing can have a config.test.local.php that overrides the $configure array without modifying the config.test.php file in the source checkout.

+++ testing/kwtest/kw_test_manager.php
@@ -299,8 +299,19 @@ class CDashTestManager extends TestManager
                                  $this->database['name'],
                                  $this->database['type']);
     $reporter->paintConfigureConnection($result);
-    $result = unlink($logfilename);
+
+    if (file_exists($logfilename))
+      {
+      // delete the log file -- result is success/failure
+      $result = unlink($logfilename);
+      }
+    else
+      {
+      // file is already not there -- equivalent to successfully deleting it
+      $result = true;
+      }
     $reporter->paintConfigureDeleteLogResult($result, $logfilename);
+
     $result = $this->_installdb4test($this->database['host'],
                                       $this->database['login'],
                                       $this->database['pwd'],
